build: align CMake config with ArcVideo (clang-cl, C++23, USE_WERROR)
- CMakeLists.txt: upgrade C++ standard from C++20 to C++23
- CMakeLists.txt: add USE_WERROR option with /WX and -Werror support
- CMakeLists.txt: align compiler options with ArcVideo pattern (/W4, clang-cl
  suppression flags, /EHsc)
- CMakePresets.json: add clang-cl compilers to windows-base
- CMakePresets.json: add ClangCL toolset to windows-vsbase
- CMakePresets.json: add USE_WERROR=true to base preset
- CMakePresets.json: add linux-base and linux-ci presets (FASTBuild, clang)
- CMakePresets.json: add -summary and targets to all build presets
- Fix -Wsign-compare warnings in color.cpp and timecodefunctions.cpp

+++ src/util/timecodefunctions.cpp
@@ -192,7 +192,7 @@ rational Timecode::timecode_to_time(std::string timecode, const rational &timeba
     static const std::regex timecode_delim("(:)|(;)");
     std::vector<std::string> timecode_split = StringUtils::split_regex(timecode, timecode_delim);
 
-    const int element_count = display == kTimecodeSeconds ? 3 : 4;
+    const size_t element_count = display == kTimecodeSeconds ? 3 : 4;
 
     // Remove excess tokens (we're only interested in HH:MM:SS.FF)
     if (timecode_split.size() > element_count) {
